requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
xgboost
catboost
dill
flask
gunicorn
PyYAML
-e .